==> notebooks/2.EDA_and_Feature_engineering.ipynb <==
import os
# перейдем в родительскую директорию чтобы использовать пакет raiflib без установки
notebook_path = os.getcwd()
os.chdir(os.path.dirname(notebook_path))

import pandas as pd
import numpy as np
from haversine import haversine, Unit
from geopy.geocoders import Nominatim

from tqdm import tqdm
from tqdm.auto import tqdm
tqdm.pandas()

df = pd.read_csv('../data/1.ipynb_Cleaning_data.csv', index_col=[0])
tdf = pd.read_csv('../data/1.ipynb_Cleaning_data_test.csv', index_col=[0])

# Добавление расстояния до столицы (Москва)

address = 'Москва'
geolocator = Nominatim(user_agent="Your_Name")
location = geolocator.reverse(f'55.750446 37.617494')

location.address

# df_locations = df.progress_apply(lambda t: geolocator.reverse(f"{t['lat']} {t['lng']}").address, axis=1)

def get_distance(coords1, coords2):
    return haversine(coords1, coords2, Unit.KILOMETERS)

def get_location(address: str) -> float:
    geolocator = Nominatim(user_agent="Your_Name")
    location = geolocator.geocode(address)
    return location.latitude, location.longitude

moscow_coords = get_location('Москва')
df['distance_to_moscow'] = df.progress_apply(
                                lambda t: get_distance((t['lat'], t['lng']), moscow_coords), 
                                axis=1
)

tdf['distance_to_moscow'] = tdf.progress_apply(
                                lambda t: get_distance((t['lat'], t['lng']), moscow_coords), 
                                axis=1
)

# Добавление расстояния до центра региона

# Сделаем Санкт-Петербург центром Ленинградской области, Москву центром московской
df['tmp_region'] = df['region'].where(df.city != 'Санкт-Петербург', 'Ленинградская область')
df['tmp_region'] = df['tmp_region'].where(df.city != 'Москва', 'Московская область')
tdf['tmp_region'] = tdf['region'].where(tdf.city != 'Санкт-Петербург', 'Ленинградская область')
tdf['tmp_region'] = tdf['tmp_region'].where(tdf.city != 'Москва', 'Московская область')

region_centers = df.groupby(['tmp_region', 'city'], as_index=False) \
                   .agg({'id':'count'}) \
                   .sort_values(by='id', ascending=False) \
                   .groupby('tmp_region').head(1) \
                   .rename(columns={
                       'id':'count',
                       'city': 'region_center'
                    })

region_centers['coords'] = region_centers.progress_apply(
                                lambda t: get_location(t['region_center']), 
                                axis=1
)
region_centers[['lat','lng']] = pd.DataFrame(region_centers['coords'].tolist(), index= region_centers.index)
region_centers = region_centers.drop(columns=['coords', 'count'])
region_centers

df = pd.merge(left=df, right=region_centers, on=['tmp_region'], how='left') \
    .rename(columns={
        'lat_x': 'lat',
        'lng_x': 'lng'
    })

df['distance_to_region_center'] = df.progress_apply(
                                lambda t: get_distance((t['lat'], t['lng']), (t['lat_y'], t['lng_y'])), 
                                axis=1
)

df = df.drop(columns=['lat_y', 'lng_y', 'region_center', 'tmp_region'])
# ----
tdf = pd.merge(left=tdf, right=region_centers, on=['tmp_region'], how='left') \
    .rename(columns={
        'lat_x': 'lat',
        'lng_x': 'lng'
    })

tdf['distance_to_region_center'] = tdf.progress_apply(
                                lambda t: get_distance((t['lat'], t['lng']), (t['lat_y'], t['lng_y'])), 
                                axis=1
)

tdf = tdf.drop(columns=['lat_y', 'lng_y', 'region_center', 'tmp_region'])

# EDA числовых признаков

# Заметим, что переменная total_square имеет распределение, близкое к лог-нормальному
df['total_square'] = df.apply(lambda t: np.log(t['total_square']), axis=1)

tdf['total_square'] = tdf.total_square.apply(np.log)

tdf.to_csv('../data/2.ipynb_EDA_and_Feature_engineering_test.csv')
df.to_csv('../data/2.ipynb_EDA_and_Feature_engineering.csv')
